+++ front/src/app/idea-memos/IdeaMemos.module.scss
